Add files via upload
The proposed project tile under UCR is “5 MW Tangling Small Hydro Electric Project in Kinnaur District of Himachal Pradesh”, which is a Hydro Power project located in village Tangling of Kinnaur district in the state of Himachal Pradesh (India). The project is an operational activity with continuous reduction of GHG, currently being applied for Registration under “Universal Carbon Registry” (UCR).